Splits Memory check out to Windows as not all metrics are available
